requirements.txt
Django~=3.2.12
djangorestframework~=3.13.1
django-filter~=21.1
pip~=22.0.2
django-environ==0.9.0
gunicorn==20.1.0
psycopg==3.0.15
psycopg2-binary==2.9.3
djoser==2.1.0